Syntax Reference

StaticLint.followinclude

StaticLint.get_parent_fexpr

StaticLint.get_path

StaticLint.is_in_fexpr

StaticLint.traverse

Main

StaticLint.followinclude — Method

followinclude(x, state)

Checks whether the arguments of a call to include can be resolved to a path. If successful it checks whether a file with that path is loaded on the server or a file exists on the disc that can be loaded. If this is successful it traverses the code associated with the loaded file.

StaticLint.traverse — Method

traverse(x, state)

Iterates across the child nodes of an EXPR in execution order (rather than storage order) calling state on each node.

StaticLint.get_path — Method

get_path(x::EXPR)

Usually called on the argument to include calls, and attempts to determine the path of the file to be included. Has limited support for joinpath calls.

StaticLint.get_parent_fexpr — Method

get_in_fexpr(x::EXPR, f)

Get the parent of x for which f(parent) == true. (isinfexpr should be called first.)

StaticLint.is_in_fexpr — Method

is_in_fexpr(x::EXPR, f)

Check whether x isa the child of an expression for which f(parent) == true.

Linting