Syntax Reference

StaticLint.addmoduletoscope!

StaticLint.check_kw_default

StaticLint.collect_hints

StaticLint.eventually_overloads

StaticLint.followinclude

StaticLint.get_parent_fexpr

StaticLint.get_path

StaticLint.initial_pass_on_exports

StaticLint.interpret_eval

StaticLint.introduces_scope

StaticLint.is_in_fexpr

StaticLint.maybe_quoted_list

StaticLint.module_safety_trip

StaticLint.new_within_struct

StaticLint.resolve_getfield

StaticLint.scope_exports

StaticLint.scopehasbinding

StaticLint.scopehasmodule

StaticLint.scopes

StaticLint.traverse

StaticLint.valofid

Main

StaticLint.followinclude — Method

followinclude(x, state)

Checks whether the arguments of a call to include can be resolved to a path. If successful it checks whether a file with that path is loaded on the server or a file exists on the disc that can be loaded. If this is successful it traverses the code associated with the loaded file.

StaticLint.traverse — Method

traverse(x, state)

Iterates across the child nodes of an EXPR in execution order (rather than storage order) calling state on each node.

StaticLint.eventually_overloads — Method

eventually_overloads(b, x, server)

StaticLint.interpret_eval — Method

interpret_eval(x::EXPR, state)

Naive attempt to interpret x as though it has been eval'ed. Lifts any bindings made within the scope of x to the toplevel and replaces (some) interpolated binding names with the value where possible.

StaticLint.maybe_quoted_list — Method

maybe_quoted_list(x::EXPR)

Try and get a list of quoted symbols from x. Return nothing if not possible.

StaticLint.initial_pass_on_exports — Method

initial_pass_on_exports(x::EXPR, server)

Export statements need to be (pseudo) evaluated each time we consider whether a variable is made available by an import statement.

StaticLint.module_safety_trip — Method

module_safety_trip(scope::Scope,  visited_scopes)

Checks whether the scope is a module and we've visited it before, otherwise adds the module to the list.

StaticLint.new_within_struct — Method

newwithinstruct(x::EXPR)

Checks whether x is a reference to new within a datatype constructor.

StaticLint.resolve_getfield — Method

resolve_getfield(x::EXPR, parent::Union{EXPR,Scope,ModuleStore,Binding}, state::State)::Bool

Given an expression of the form parent.x try to resolve x. The method called with parent::EXPR resolves the reference for parent, other methods then check whether the Binding/Scope/ModuleStore to which parent points has a field matching x.

StaticLint.scope_exports — Method

scope_exports(scope::Scope, name::String)

Does the scope export a variable called name?

StaticLint.valofid — Method

valofid(x)

Returns the string value of an expression for which isidentifier is true, i.e. handles NONSTDIDENTIFIERs.

StaticLint.addmoduletoscope! — Method

addmoduletoscope!(s, m, [mname::Symbol])

Adds module m to the list of used modules in scope s.

StaticLint.introduces_scope — Method

introduces_scope(x::EXPR, state)

Does this expression introduce a new scope?

StaticLint.scopehasbinding — Method

scopehasbinding(s::Scope, n::String)

Checks whether s has a binding for variable named n.

StaticLint.scopehasmodule — Method

scopehasmodule(s::Scope, mname::Symbol)::Bool

Checks whether the module mname has been usinged in s.

StaticLint.scopes — Method

scopes(x::EXPR, state)

Called when traversing the syntax tree and handles the association of scopes with expressions. On the first pass this will add scopes as necessary, on following passes it empties it.

StaticLint.get_path — Method

get_path(x::EXPR)

Usually called on the argument to include calls, and attempts to determine the path of the file to be included. Has limited support for joinpath calls.

StaticLint.get_parent_fexpr — Method

get_in_fexpr(x::EXPR, f)

Get the parent of x for which f(parent) == true. (isinfexpr should be called first.)

StaticLint.is_in_fexpr — Method

is_in_fexpr(x::EXPR, f)

Check whether x isa the child of an expression for which f(parent) == true.

Linting

StaticLint.check_kw_default — Method

check_kw_default(x::EXPR, server)

Check that the default value matches the type for keyword arguments. Following types are checked: String, Symbol, Int, Char, Bool, Float32, Float64, UInt8, UInt16, UInt32, UInt64, UInt128.

StaticLint.collect_hints — Function

collect_hints(x::EXPR, server, missingrefs = :all, isquoted = false, errs = Tuple{Int,EXPR}[], pos = 0)

Collect hints and errors from an expression. missingrefs = (:none, :id, :all) determines whether unresolved identifiers are marked, the :all option will mark identifiers used in getfield calls."